升级测试框架,过滤压测开始时候的请求响应,增加测试开关.

+++ src/main/groovy/com/funtester/frame/execute/FixedQpsConcurrent.java
@@ -47,7 +47,7 @@ public class FixedQpsConcurrent extends SourceCode {
     /**
      * 用于记录所有请求时间
      */
-    public static Vector<Integer> allTimes = new Vector<>();
+    public static Vector<Short> allTimes = new Vector<>();
 
     /**
      * 开始时间
